# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/login":{"post":{"tags":["login"],"summary":"Login","operationId":"login_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users":{"get":{"tags":["users"],"summary":"Get Users","operationId":"get_users_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserResponse"},"type":"array","title":"Response Get Users Users Get"}}}}}},"post":{"tags":["users"],"summary":"Create User","operationId":"create_user_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{id}":{"get":{"tags":["users"],"summary":"Get User","operationId":"get_user_users__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/email/{email}":{"get":{"tags":["users"],"summary":"Get User By Email","operationId":"get_user_by_email_users_email__email__get","parameters":[{"name":"email","in":"path","required":true,"schema":{"type":"string","title":"Email"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/articles":{"get":{"tags":["articles"],"summary":"Get Articles","operationId":"get_articles_articles_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ArticleResponse"},"type":"array","title":"Response Get Articles Articles Get"}}}}}},"post":{"tags":["articles"],"summary":"Create Article","operationId":"create_article_articles_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/articles/{id}":{"get":{"tags":["articles"],"summary":"Get Article","operationId":"get_article_articles__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reviews":{"get":{"tags":["reviews"],"summary":"Get Reviews","operationId":"get_reviews_reviews_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ReviewResponse"},"type":"array","title":"Response Get Reviews Reviews Get"}}}}}},"put":{"tags":["reviews"],"summary":"Update Review","operationId":"update_review_reviews_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["reviews"],"summary":"Create Review","operationId":"create_review_reviews_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reviews/{id}":{"get":{"tags":["reviews"],"summary":"Get Review","operationId":"get_review_reviews__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advent_calendars":{"get":{"tags":["advent_calendars"],"summary":"Get Advent Calendars","operationId":"get_advent_calendars_advent_calendars_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AdventCalendarResponse"},"type":"array","title":"Response Get Advent Calendars Advent Calendars Get"}}}}}},"post":{"tags":["advent_calendars"],"summary":"Create Advent Calendar","operationId":"create_advent_calendar_advent_calendars_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdventCalendarCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdventCalendarResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advent_calendars/year/{year}}":{"get":{"tags":["advent_calendars"],"summary":"Get Advent Calendars By Year","operationId":"get_advent_calendars_by_year_advent_calendars_year__year___get","parameters":[{"name":"year","in":"path","required":true,"schema":{"type":"integer","title":"Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdventCalendarResponse"},"title":"Response Get Advent Calendars By Year Advent Calendars Year  Year   Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advent_calendars/year/{year}/{name}}":{"get":{"tags":["advent_calendars"],"summary":"Get Advent Calendars By Year And Name","operationId":"get_advent_calendars_by_year_and_name_advent_calendars_year__year___name___get","parameters":[{"name":"year","in":"path","required":true,"schema":{"type":"integer","title":"Year"}},{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdventCalendarResponse"},"title":"Response Get Advent Calendars By Year And Name Advent Calendars Year  Year   Name   Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advent_calendars/{id}":{"get":{"tags":["advent_calendars"],"summary":"Get Advent Calendar","operationId":"get_advent_calendar_advent_calendars__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdventCalendarResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advent_calendar_articles":{"get":{"tags":["advent_calendar_articles"],"summary":"Get Advent Calendar Articles","operationId":"get_advent_calendar_articles_advent_calendar_articles_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AdventCalendarArticleResponse"},"type":"array","title":"Response Get Advent Calendar Articles Advent Calendar Articles Get"}}}}}},"post":{"tags":["advent_calendar_articles"],"summary":"Create Advent Calendar Article","operationId":"create_advent_calendar_article_advent_calendar_articles_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdventCalendarArticleCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdventCalendarArticleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advent_calendar_articles/ac/{id}":{"get":{"tags":["advent_calendar_articles"],"summary":"Get Advent Calendar Articles By Ac","operationId":"get_advent_calendar_articles_by_ac_advent_calendar_articles_ac__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdventCalendarArticleResponse"},"title":"Response Get Advent Calendar Articles By Ac Advent Calendar Articles Ac  Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advent_calendar_articles/ac/{id}/{day}":{"get":{"tags":["advent_calendar_articles"],"summary":"Get Advent Calendar Articles By Ac And Day","operationId":"get_advent_calendar_articles_by_ac_and_day_advent_calendar_articles_ac__id___day__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}},{"name":"day","in":"path","required":true,"schema":{"type":"integer","title":"Day"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdventCalendarArticleResponse"},"title":"Response Get Advent Calendar Articles By Ac And Day Advent Calendar Articles Ac  Id   Day  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/advent_calendar_articles/{id}":{"get":{"tags":["advent_calendar_articles"],"summary":"Get Advent Calendar Article","operationId":"get_advent_calendar_article_advent_calendar_articles__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdventCalendarArticleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AdventCalendarArticleCreate":{"properties":{"advent_calendar_id":{"type":"string","title":"Advent Calendar Id"},"day":{"type":"integer","title":"Day"},"title":{"type":"string","title":"Title"},"article_id":{"type":"string","title":"Article Id"},"author_id":{"type":"string","title":"Author Id"}},"type":"object","required":["advent_calendar_id","day","title","article_id","author_id"],"title":"AdventCalendarArticleCreate"},"AdventCalendarArticleResponse":{"properties":{"id":{"type":"string","title":"Id"},"advent_calendar_id":{"type":"string","title":"Advent Calendar Id"},"day":{"type":"integer","title":"Day"},"title":{"type":"string","title":"Title"},"article_id":{"type":"string","title":"Article Id"},"author_id":{"type":"string","title":"Author Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"}},"type":"object","required":["id","advent_calendar_id","day","title","article_id","author_id","created_at","updated_at","deleted_at"],"title":"AdventCalendarArticleResponse"},"AdventCalendarCreate":{"properties":{"year":{"type":"integer","title":"Year"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"author_id":{"type":"string","title":"Author Id"}},"type":"object","required":["year","title","description","author_id"],"title":"AdventCalendarCreate"},"AdventCalendarResponse":{"properties":{"id":{"type":"string","title":"Id"},"year":{"type":"integer","title":"Year"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"author_id":{"type":"string","title":"Author Id"}},"type":"object","required":["id","year","title","description","author_id"],"title":"AdventCalendarResponse"},"ArticleCreate":{"properties":{"title":{"type":"string","maxLength":50,"title":"Title","example":"title"},"description":{"type":"string","maxLength":50,"title":"Description","example":"content"},"author_id":{"type":"string","maxLength":50,"title":"Author Id","example":"author_id"},"path":{"type":"string","maxLength":50,"title":"Path","example":"path"},"filename":{"type":"string","maxLength":50,"title":"Filename","example":"filename"},"allow_external":{"type":"boolean","title":"Allow External","example":true},"show_from":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Show From","example":"2023-11-12 00:00:00"},"show_until":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Show Until","example":"2099-12-31 23:59:59"},"review_ok":{"type":"boolean","title":"Review Ok","example":true}},"type":"object","required":["title","description","author_id","path","filename","allow_external","show_from","show_until","review_ok"],"title":"ArticleCreate"},"ArticleResponse":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"author_id":{"type":"string","title":"Author Id"},"author":{"$ref":"#/components/schemas/User"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"},"path":{"type":"string","title":"Path"},"filename":{"type":"string","title":"Filename"},"allow_external":{"type":"boolean","title":"Allow External"},"show_from":{"type":"string","format":"date-time","title":"Show From"},"show_until":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Show Until"},"review_ok":{"type":"boolean","title":"Review Ok"},"tags":{"items":{},"type":"array","title":"Tags"},"comments":{"items":{},"type":"array","title":"Comments"},"likes":{"items":{},"type":"array","title":"Likes"},"reviews":{"items":{},"type":"array","title":"Reviews"}},"type":"object","required":["id","title","description","author_id","author","created_at","updated_at","deleted_at","path","filename","allow_external","show_from","show_until","review_ok","tags","comments","likes","reviews"],"title":"ArticleResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","maxLength":50,"minLength":3,"title":"Email","example":"test@example.com"},"password":{"type":"string","maxLength":50,"minLength":8,"title":"Password","example":"password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"LoginResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"}},"type":"object","required":["access_token","id","name","email"],"title":"LoginResponse"},"ReviewCreate":{"properties":{"comment":{"type":"string","title":"Comment"},"reviewer_id":{"type":"string","title":"Reviewer Id"},"article_id":{"type":"string","title":"Article Id"}},"type":"object","required":["comment","reviewer_id","article_id"],"title":"ReviewCreate"},"ReviewResponse":{"properties":{"id":{"type":"string","title":"Id"},"comment":{"type":"string","title":"Comment"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"},"reviewer_id":{"type":"string","title":"Reviewer Id"},"article_id":{"type":"string","title":"Article Id"},"approved":{"type":"boolean","title":"Approved"}},"type":"object","required":["id","comment","created_at","updated_at","deleted_at","reviewer_id","article_id","approved"],"title":"ReviewResponse"},"ReviewUpdate":{"properties":{"id":{"type":"string","title":"Id"},"comment":{"type":"string","title":"Comment"},"approved":{"type":"boolean","title":"Approved"}},"type":"object","required":["id","comment","approved"],"title":"ReviewUpdate"},"User":{"properties":{"id":{"type":"string","maxLength":50,"title":"Id","example":"id"},"name":{"type":"string","maxLength":50,"title":"Name","example":"name"},"email":{"type":"string","maxLength":50,"title":"Email","example":"email"},"password":{"type":"string","maxLength":50,"title":"Password","example":"password"},"role":{"type":"string","maxLength":50,"title":"Role","example":"role"},"created_at":{"type":"string","title":"Created At","example":"2023-11-12 00:00:00"},"updated_at":{"type":"string","title":"Updated At","example":"2023-11-12 00:00:00"},"deleted_at":{"type":"string","title":"Deleted At","example":"2023-11-12 00:00:00"}},"type":"object","required":["id","name","email","password","role","created_at","updated_at","deleted_at"],"title":"User"},"UserCreate":{"properties":{"email":{"type":"string","maxLength":50,"minLength":3,"title":"Email","example":"test@example.com"},"password":{"type":"string","maxLength":50,"minLength":8,"title":"Password","example":"password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"UserResponse":{"properties":{"id":{"type":"string","title":"Id"},"email":{"type":"string","title":"Email"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","email","name"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}